+++ js/ui/map.js
@@ -24,7 +24,6 @@ var Map = module.exports = function(options) {
 
     this.animationLoop = new AnimationLoop();
     this.transform = new Transform(options.minZoom, options.maxZoom);
-    this.hash = options.hash && new Hash(this);
 
     if (options.maxBounds) {
         var b = LatLngBounds.convert(options.maxBounds);
@@ -36,6 +35,7 @@ var Map = module.exports = function(options) {
         '_forwardStyleEvent',
         '_forwardSourceEvent',
         '_forwardTileEvent',
+        '_onStyleLoad',
         '_onStyleChange',
         '_onSourceAdd',
         '_onSourceRemove',
@@ -49,16 +49,19 @@ var Map = module.exports = function(options) {
 
     this.handlers = options.interactive && new Handlers(this);
 
-     // don't set position from options if set through hash
-    if (!this.hash || !this.hash.onhash()) {
+    this._hash = options.hash && (new Hash()).addTo(this);
+    // don't set position from options if set through hash
+    if (!this._hash || !this._hash._onHashChange()) {
         this.setView(options.center, options.zoom, options.bearing);
     }
 
     this.sources = {};
     this.stacks = {};
+    this._classes = {};
 
     this.resize();
 
+    if (options.classes) this.setClasses(options.classes);
     if (options.style) this.setStyle(options.style);
     if (options.attributionControl) this.addControl(new Attribution());
 };
@@ -130,6 +133,34 @@ util.extend(Map.prototype, {
     getZoom() { return this.transform.zoom; },
     getBearing() { return this.transform.bearing; },
 
+    addClass(klass, options) {
+        if (this._classes[klass]) return;
+        this._classes[klass] = true;
+        if (this.style) this.style._cascadeClasses(this._classes, options);
+    },
+
+    removeClass(klass, options) {
+        if (!this._classes[klass]) return;
+        delete this._classes[klass];
+        if (this.style) this.style._cascadeClasses(this._classes, options);
+    },
+
+    setClasses(klasses, options) {
+        this._classes = {};
+        for (var i = 0; i < klasses.length; i++) {
+            this._classes[klasses[i]] = true;
+        }
+        if (this.style) this.style._cascadeClasses(this._classes, options);
+    },
+
+    hasClass(klass) {
+        return !!this._classes[klass];
+    },
+
+    getClasses() {
+        return Object.keys(this._classes);
+    },
+
     // Detect the map's new width and height and resize it.
     resize() {
         var width = 0, height = 0;
@@ -175,7 +206,7 @@ util.extend(Map.prototype, {
     setStyle(style) {
         if (this.style) {
             this.style
-                .off('load', this._forwardStyleEvent)
+                .off('load', this._onStyleLoad)
                 .off('error', this._forwardStyleEvent)
                 .off('change', this._onStyleChange)
                 .off('source.add', this._onSourceAdd)
@@ -200,7 +231,7 @@ util.extend(Map.prototype, {
         }
 
         this.style
-            .on('load', this._forwardStyleEvent)
+            .on('load', this._onStyleLoad)
             .on('error', this._forwardStyleEvent)
             .on('change', this._onStyleChange)
             .on('source.add', this._onSourceAdd)
@@ -304,6 +335,11 @@ util.extend(Map.prototype, {
 
         this.fire('render');
 
+        if (this.loaded() && !this._loaded) {
+            this._loaded = true;
+            this.fire('load');
+        }
+
         this._frameId = null;
 
         if (!this.animationLoop.stopped()) {
@@ -318,6 +354,7 @@ util.extend(Map.prototype, {
     },
 
     remove() {
+        if (this._hash) this._hash.remove();
         browser.cancelFrame(this._frameId);
         clearTimeout(this._sourcesDirtyTimeout);
         this.setStyle(null);
@@ -340,6 +377,11 @@ util.extend(Map.prototype, {
 
     _forwardTileEvent(e) {
         this.fire(e.type, util.extend({style: e.target}, e));
+    },
+
+    _onStyleLoad(e) {
+        this.style._cascade(this._classes, {transition: false});
+        this._forwardStyleEvent(e);
     },
 
     _onStyleChange(e) {