+++ CHANGELOG.md
@@ -5,6 +5,9 @@ An in-progress version being developed in the `mb-pages` branch.
 #### Breaking
 
 * HTTP_URL and HTTPS_URL config options must no longer include a `/v4` path prefix.
+* `addClass`, `removeClass`, `setClasses`, `hasClass`, and `getClasses` are now methods
+  on Map.
+* `Style#cascade` is now private, pending a public style mutation API (#755).
 
 ## 0.5.2 (Jan 07 2015)